+++ src/components/Footer/Footer.tsx
@@ -1,5 +1,5 @@
 import React from 'react'
-import s from '../css/Footer.module.css'
+import s from './Footer.module.css'
