+++ src/components/Profile/MyPosts/Post/Post.module.css
@@ -1,0 +1,7 @@
+.post img {
+width: 50px;
+    border-radius: 50%;
+}
+button {
+    display: inline;
+}
