--- src/css/Footer.module.css
@@ -1,3 +1,0 @@
-.footer {
-    grid-area: f
-}
